Move urlutils to pkg2
