+++ GitlabProject.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/asciich/asciichgolangpublic/logging"
 	"github.com/asciich/asciichgolangpublic/parameteroptions/authenticationoptions"
+	"github.com/asciich/asciichgolangpublic/pkg/urlsutils"
 	"github.com/asciich/asciichgolangpublic/pkg/versionutils"
 	"github.com/asciich/asciichgolangpublic/tracederrors"
-	"github.com/asciich/asciichgolangpublic/urlsutils"
 	gitlab "gitlab.com/gitlab-org/api/client-go"
 )
